Bugfix SqlServer when using schema names
